+++ Observer/BounceIp.php
@@ -32,6 +32,7 @@ use Magento\Framework\Event\Observer;
 use Magento\Framework\Event\ObserverInterface;
 use CrowdSec\Engine\Helper\Data as Helper;
 use CrowdSec\Engine\CapiEngine\Remediation;
+use CrowdSec\RemediationEngine\Constants as RemediationConstants;
 use Magento\Framework\HTTP\PhpEnvironment\Response;
 use CrowdSec\Engine\Constants;
 use Magento\Store\Model\StoreManagerInterface;
@@ -104,7 +105,8 @@ class BounceIp implements ObserverInterface
             }
 
             $ip = $this->helper->getRealIp();
-            $remediation = $this->remediation->getIpRemediation($ip);
+            $remediationData = $this->remediation->getIpRemediation($ip);
+            $remediation = $remediationData[RemediationConstants::REMEDIATION_KEY]??Constants::REMEDIATION_BYPASS;
             if ($remediation === Constants::REMEDIATION_BAN) {
                 /**
                  * @var $response Response
@@ -122,6 +124,9 @@ class BounceIp implements ObserverInterface
                 }
 
                 $response->setBody($content)->setStatusCode(Http::STATUS_CODE_403);
+                $this->remediation->updateMetricsOriginsCount(
+                    $remediationData[RemediationConstants::ORIGIN_KEY], $remediation
+                );
             }
         } catch (\Exception $e) {
             $this->helper->getLogger()->error('Technical error while bouncing ip', ['message' => $e->getMessage()]);